pyproject.toml
[project]
name = "galileo-onboarding-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "galileo[openai]>=1.39.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-openai>=0.3.35",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
